Maybe this will fix the 0% issue?

+++ GameLibrary/sources/Modifier.cpp
@@ -104,5 +104,5 @@ QString Modifier::ToString()
  */
 QString Modifier::OperationToString()
 {
-  return application_type_ == ModifierOperation::Multiplicative ? QString::number((abs(amount_) - 1.0) * 100) + "%" : QString::number(abs(amount_));
+  return application_type_ == ModifierOperation::Multiplicative ? QString::number(static_cast<float>((abs(amount_) - 1.0) * 100.0)) + "%" : QString::number(abs(amount_));
 }
